Refactor: 바로등기 폴더구조 변경
