+++ services/ggi-web/app/deunggi/components/ContentBottom/SearchListBox/SearchList/style.ts
