+++ services/ggi-web/app/deunggi/components/ContentBottom/SearchListBox/style.ts
