+++ services/ggi-web/app/deunggi/components/ContentBottom/ViewListBox/ViewList/style.ts
