+++ services/ggi-web/app/deunggi/components/ContentBottom/ViewListBox/style.ts
