+++ services/ggi-web/app/deunggi/components/ContentBottom/index.tsx
