+++ services/ggi-web/app/deunggi/components/ContentTop/DeunggiInfo/index.tsx
