+++ services/ggi-web/app/deunggi/components/ContentTop/DeunggiInfo/style.ts
