+++ services/ggi-web/app/deunggi/components/ContentTop/StrokeButton/index.tsx
