+++ services/ggi-web/app/deunggi/components/ContentTop/StrokeButton/style.ts
