+++ services/ggi-web/app/deunggi/page.tsx
@@ -1,6 +1,6 @@
 'use client'
-import DeunggiSection from "./components/deunggi/DeunggiSection";
-import Header from "./components/header/headerComponents";
+import DeunggiSection from './components/DeunggiSection'
+import Header from './components/header/headerComponents'
 
 export default function Deunggi() {
   return (
